test: fix relative paths in specs that moved one level deeper (#26)
new_call_collector and method_type_resolver: support helper require
(../../ -> ../../../). steep_bridge: dummy fixture reads (../../dummy ->
../../../dummy). analyzer_spec was untouched (stayed at top level).
